requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
plotly
openai
python-dotenv
requests
vaderSentiment
newsapi-python
pycoingecko
